+++ src/Redmine/Api/User.php
@@ -118,6 +118,9 @@ class User extends AbstractApi
     /**
      * Returns an array of users with login/id pairs.
      *
+     * @deprecated v2.7.0 Use listLogins() instead.
+     * @see User::listLogins()
+     *
      * @param bool  $forceUpdate to force the update of the users var
      * @param array $params      to allow offset/limit (and more) to be passed
      *
@@ -125,6 +128,8 @@ class User extends AbstractApi
      */
     public function listing($forceUpdate = false, array $params = [])
     {
+        @trigger_error('`' . __METHOD__ . '()` is deprecated since v2.7.0, use `' . __CLASS__ . '::listLogins()` instead.', E_USER_DEPRECATED);
+
         if (empty($this->users) || $forceUpdate) {
             $this->users = $this->list($params);
         }